Fix Z-Image grain: pytorch-compatible VAE GroupNorm + fp32 adaLN + mflux schedule
The speckle/'mosaic' grain (vs the clean mflux reference on the same Q4 weights)
was the VAE GroupNorm using MLX's default INTERLEAVED channel grouping instead of
PyTorch-contiguous groups — every decoder norm computed the wrong per-pixel
mean/variance. Setting pytorchCompatible:true on the groupNorm helper (routes
through fp32-internal mx.fast.layer_norm) clears it across all VAE norms.

Also matched mflux to close the remaining gap:
- adaLN keeps the timestep embedding in fp32 through the modulation Linear (block
  + final layer), so the residual stream is fp32 (bf16xfp32 promotion) not bf16.
- Sampler schedule adds shift_terminal=0.02 + exponential shift (exp(1)), so the
  sigmas are bit-identical to mflux (…0.325, 0.02, 0); the model is no longer
  evaluated below its calibrated range.
- Caption padded to a multiple of 32 (already in a prior commit) + VAE fp32 norms.

+++ Sources/ZImageMLX/ZImageDenoiser.swift
@@ -82,12 +82,14 @@ final class ZImageFinalLayer: Module {
         super.init()
     }
     func callAsFunction(_ x: MLXArray, timeEmb: MLXArray) -> MLXArray {
-        let scale = expandedDimensions(adaLNModulation[0](silu(timeEmb.asType(x.dtype))), axis: 1)
+        // Keep the timestep embedding fp32 through silu + the adaLN Linear (matches mflux), so the
+        // scale-only modulation and the final projection stay fp32.
+        let scale = expandedDimensions(adaLNModulation[0](silu(timeEmb)), axis: 1)
         // Parameter-free LayerNorm (no affine) computed in fp32 for stability, then scale-only AdaLN.
         let xf = x.asType(.float32)
         let mean = xf.mean(axis: -1, keepDims: true)
         let variance = xf.variance(axis: -1, keepDims: true)
-        let normed = ((xf - mean) * rsqrt(variance + eps)).asType(x.dtype)
+        let normed = (xf - mean) * rsqrt(variance + eps)
         return linear(normed * (1 + scale))
     }
 }
